Cleaned up services and exception handling...

+++ book-service/src/main/java/com/pungu/store/book_service/services/BookServiceImpl.java
@@ -4,9 +4,12 @@ import com.pungu.store.book_service.clients.AuthorClient;
 import com.pungu.store.book_service.dtos.BookRequest;
 import com.pungu.store.book_service.dtos.BookResponse;
 import com.pungu.store.book_service.entities.Book;
+import com.pungu.store.book_service.exceptions.BookAlreadyExistsException;
+import com.pungu.store.book_service.exceptions.BookNotFoundException;
 import com.pungu.store.book_service.repositories.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,51 +24,6 @@ public class BookServiceImpl implements BookService {
     private final AuthorClient authorClient;
 
     /**
-     * Creates a new book entry in the system.
-     *
-     * @param bookRequest Request DTO containing book details.
-     * @return BookResponse DTO with the created book's details.
-     */
-    @Override
-    @Transactional
-    public BookResponse createBook(BookRequest bookRequest) {
-        Long authorId = bookRequest.getAuthorName() != null ? authorClient.getAuthorIdByName(bookRequest.getAuthorName()) : bookRequest.getAuthorId();
-
-        Book book = Book.builder()
-                .title(bookRequest.getTitle())
-                .authorId(authorId)
-                .description(bookRequest.getDescription())
-                .genre(bookRequest.getGenre())
-                .language(bookRequest.getLanguage())
-                .publicationDate(bookRequest.getPublicationDate())
-                .ebookUrl(bookRequest.getEbookUrl())
-                .coverImageUrl(bookRequest.getCoverImageUrl())
-                .availableForReading(bookRequest.getEbookUrl() != null)
-                .availableForDownload(bookRequest.getEbookUrl() != null)
-                .build();
-
-        return entityToResponse(bookRepository.save(book));
-    }
-
-    /**
-     * Retrieves a book by its ID.
-     *
-     * @param bookId ID of the book.
-     * @return BookResponse with book details.
-     */
-    @Override
-    public BookResponse getBookById(Long bookId) {
-        Book book = bookRepository.findById(bookId)
-                .orElseThrow(() -> new RuntimeException("Book not found"));
-        return entityToResponse(book);
-    }
-
-    @Override
-    public List<BookResponse> getAllBookByAuthorId(Long authorId, Sort sort) {
-        return bookRepository.findByAuthorId(authorId, sort).stream().map(this::entityToResponse).collect(Collectors.toList());
-    }
-
-    /**
      * Returns all books in the database.
      *
      * @return List of BookResponse DTOs.
@@ -78,6 +36,60 @@ public class BookServiceImpl implements BookService {
                 .collect(Collectors.toList());
     }
 
+
+    /**
+     * Creates a new book entry in the system.
+     *
+     * @param bookRequest Request DTO containing book details.
+     * @return BookResponse DTO with the created book's details.
+     */
+    @Override
+    @Transactional
+    @PreAuthorize("hasRole('ADMIN')")
+    public BookResponse createBook(BookRequest bookRequest) {
+        if (bookRepository.existsByTitleIgnoreCase(bookRequest.getTitle())) {
+            throw new BookAlreadyExistsException("Book with this title already exists.");
+        }
+
+        Book book = Book.builder()
+                .title(bookRequest.getTitle())
+                .authorId(bookRequest.getAuthorId())
+                .authorName(getAuthorName(bookRequest.getAuthorId()))
+                .description(bookRequest.getDescription())
+                .genre(bookRequest.getGenre())
+                .language(bookRequest.getLanguage())
+                .publicationDate(bookRequest.getPublicationDate())
+                .coverImageUrl(bookRequest.getCoverImageUrl())
+                .build();
+
+        return entityToResponse(bookRepository.save(book));
+    }
+
+
+    /**
+     * Returns all books for a particular author in the database.
+     *
+     * @return List of BookResponse DTOs.
+     */
+    @Override
+    public List<BookResponse> getAllBookByAuthorId(Long authorId, Sort sort) {
+        return bookRepository.findByAuthorId(authorId, sort).stream().map(this::entityToResponse).collect(Collectors.toList());
+    }
+
+    /**
+     * Retrieves a book by its ID.
+     *
+     * @param bookId ID of the book.
+     * @return BookResponse with book details.
+     */
+    @Override
+    public BookResponse getBookById(Long bookId) {
+        Book book = bookRepository.findById(bookId)
+                .orElseThrow(() -> new BookNotFoundException("There is no book found for this book id " + bookId));
+        return entityToResponse(book);
+    }
+
+
     /**
      * Updates an existing book.
      *
@@ -87,28 +99,25 @@ public class BookServiceImpl implements BookService {
      */
     @Override
     @Transactional
+    @PreAuthorize("hasRole('ADMIN')")
     public BookResponse updateBook(Long bookId, BookRequest bookRequest) {
-        if (!bookRepository.existsById(bookId)) {
-            throw new RuntimeException("Book not found");
-        }
+        Book book = bookRepository.findById(bookId)
+                .orElseThrow(() -> new BookNotFoundException(
+                        "There is no book found for this book id " + bookId));
 
-        Long authorId = bookRequest.getAuthorName() != null ? authorClient.getAuthorIdByName(bookRequest.getAuthorName()) : bookRequest.getAuthorId();
+        book.setTitle(bookRequest.getTitle());
+        book.setAuthorId(bookRequest.getAuthorId());
+        book.setAuthorName(getAuthorName(bookRequest.getAuthorId()));
+        book.setDescription(bookRequest.getDescription());
+        book.setGenre(bookRequest.getGenre());
+        book.setLanguage(bookRequest.getLanguage());
+        book.setPublicationDate(bookRequest.getPublicationDate());
+        book.setCoverImageUrl(bookRequest.getCoverImageUrl());
 
-        Book updatedBook = Book.builder()
-                .title(bookRequest.getTitle())
-                .authorId(authorId)
-                .description(bookRequest.getDescription())
-                .genre(bookRequest.getGenre())
-                .language(bookRequest.getLanguage())
-                .publicationDate(bookRequest.getPublicationDate())
-                .ebookUrl(bookRequest.getEbookUrl())
-                .coverImageUrl(bookRequest.getCoverImageUrl())
-                .availableForReading(bookRequest.getEbookUrl() != null)
-                .availableForDownload(bookRequest.getEbookUrl() != null)
-                .build();
-
-        return entityToResponse(bookRepository.save(updatedBook));
+        Book saved = bookRepository.save(book);
+        return entityToResponse(saved);
     }
+
 
     /**
      * Deletes a book by its ID.
@@ -116,9 +125,12 @@ public class BookServiceImpl implements BookService {
      * @param bookId ID of the book to delete.
      */
     @Override
+    @Transactional
+    @PreAuthorize("hasRole('ADMIN')")
     public void deleteBook(Long bookId) {
-        Book book = bookRepository.findById(bookId)
-                .orElseThrow(() -> new RuntimeException("Book not found"));
+        if (!bookRepository.existsById(bookId)) {
+            throw new BookNotFoundException("There is no book found for this book id " + bookId);
+        }
         bookRepository.deleteById(bookId);
     }
 
@@ -130,24 +142,27 @@ public class BookServiceImpl implements BookService {
      * @return BookResponse with all book details.
      */
     private BookResponse entityToResponse(Book book) {
-        String authorName = book.getAuthorId() != null
-                ? authorClient.getAuthorNameById(book.getAuthorId())
-                : "Unknown Author";
-        long authorId = book.getAuthorId() != null ? book.getAuthorId() : -1;
-
         return BookResponse.builder()
                 .bookId(book.getBookId())
                 .title(book.getTitle())
-                .authorName(authorName)
-                .authorId(authorId)
+                .authorId(book.getAuthorId())
+                .authorName(book.getAuthorName())
                 .description(book.getDescription())
                 .genre(book.getGenre())
-                .coverImageUrl(book.getCoverImageUrl())
-                .publicationDate(book.getPublicationDate())
-                .availableForReading(book.isAvailableForReading())
-                .availableForDownload(book.isAvailableForDownload())
                 .language(book.getLanguage())
+                .publicationDate(book.getPublicationDate())
+                .coverImageUrl(book.getCoverImageUrl())
                 .build();
+    }
+
+
+    private String getAuthorName(Long authorId) {
+        String fetch_author_name = "Unknown Author";
+
+        if (authorId != null) {
+            fetch_author_name = authorClient.getAuthorNameById(authorId);
+        }
+        return fetch_author_name;
     }
 
 }
